refactor(TrailingSL): split TrailingSL and strategy inheriting from it into dedicated files
